pyproject.toml
[tool.poetry]
name = "musicbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
yt-dlp = "^2023.2.17"
pillow = "^9.4.0"
discord-py = {version = "^2.2.2", extras = ["voice"]}
colorlog = "^6.7.0"
quantumrandom = "^1.9.0"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
